Moved `widget.css` to `src` so it's no longer included as a static resource
