+++ widgets/blueprintSources/src/widget.tsx
@@ -1,8 +1,8 @@
 // @ts-nocheck File not migrated fully to TS
 
-import '../widget.css';
 import Actions from './actions';
 import BlueprintSources from './BlueprintSources';
+import './widget.css';
 
 Stage.defineWidget({
     id: 'blueprintSources',
