+++ widgets/blueprints/src/widget.tsx
@@ -1,9 +1,9 @@
 // @ts-nocheck File not migrated fully to TS
-import { join } from 'lodash';
-import '../widget.css';
+import {join} from 'lodash';
 import BlueprintsLabelFilter from './BlueprintsLabelFilter';
 import BlueprintsList from './BlueprintsList';
-import type { BlueprintsWidgetConfiguration } from './types';
+import type {BlueprintsWidgetConfiguration} from './types';
+import './widget.css';
 
 const t = Stage.Utils.getT('widgets.blueprints');
 const tCatalogConfiguration = Stage.Utils.getT('widgets.blueprintCatalog.configuration');
