+++ widgets/events/src/widget.css
