+++ widgets/labels/src/widget.css
