+++ widgets/sitesMap/src/widget.css
