+++ widgets/topology/src/widget.tsx
@@ -1,10 +1,10 @@
 // @ts-nocheck File not migrated fully to TS
 
-import { castArray } from 'lodash';
-import '../widget.css';
+import {castArray} from 'lodash';
 import DataFetcher from './DataFetcher';
-import { createBaseTopology } from './DataProcessor';
+import {createBaseTopology} from './DataProcessor';
 import Topology from './Topology';
+import './widget.css';
 
 Stage.defineWidget({
     id: 'topology',
